Consolidate scalems.radical.runtime.
Move most of `scalems.radical.__init__.py` into the `runtime` submodule.

Reduce the exposure of `scalems.radical.runtime` details, and simplify
the `scalems.radical` module interfaces.

Consolidate the `_set_configuration` and `configuration` hooks used by
`scalems.invocation.run()`.
